+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [2025-07-13]
+- 新增：demo`Python/demo`文件夹，用于存放笔记当中使用的示例
+- 新增：类`Python/begin/类.md`，介绍了Python中的类
+
 ## [2025-07-08]
 - 创建`Python/syntactic_sugar`文件夹，用于储存有关语法糖的笔记
 - 新增：f-字符串`Python/syntactic_sugar/f-字符串.md`介绍Python中有关f-字符串的内容